fixed deque

+++ deque/deque.cpp
@@ -153,7 +153,7 @@ namespace stack_emu {
 	void deque<T>::resize(size_t v) {
 		size_t prev_sz = sz;
 		reserve_(v);
-		for (size_t i = begin + prev_sz; i < begin + sz; i++) {
+		for (size_t i = (size_t) (begin + prev_sz); i < (size_t) (begin + sz); i++) {
 			data[i] = T();
 		}
 	}
